Add some style and remove some old files
