+++ src/App.css
@@ -43,6 +43,32 @@ h1 {
 .w-50 {
   width: 50%
 }
+.game-image,
+.game-image-option {
+  height: 250px;
+}
+.game-image-option {
+  border: 1px solid gray;
+  border-radius: 5px;
+  padding: 5px;
+  margin-right: 10px;
+}
+
+.game-image-option:hover {
+  cursor: pointer;
+  background: green;
+}
+
+.scorebar {
+  list-style: none;
+  padding: 0;
+  border: 1px solid black;
+  border-radius: 5px;
+}
+.scorebar li {
+  display: inline-block;
+  padding: 10px;
+}
 /* FRONTPAGE */
 .frontpage-wrapper {
   text-align: center;
