--- src/components/GameOne.css
@@ -1,15 +1,0 @@
-.game-container{
-    display:flex;
-}
-
-.score{
-    font-size: 30px;
-    width: 30%;
-}
-
-.game{
-    font-size: 30px;
-    width: 70%;
-}
-
-
