+++ src/components/Scorebar.js
@@ -2,26 +2,18 @@ import React from 'react'
 import { connect } from 'react-redux';
 import { resetScore } from '../actions/scoreActions'
 
-class Scorebar extends React.Component{
-    render(){
-        const { score, streak, totalQuestions, level, successRate} = this.props.scoreState;
+class Scorebar extends React.Component {
+    render() {
+        const { score, streak, totalQuestions, level, successRate } = this.props.scoreState;
         return (
-            <div className="score">
-                <p>
-                    Level {level}
-                </p>
-                <p>
-                    No. of Questions Answered {totalQuestions}
-                </p>
-                <p>
-                    No.of Correct Answers {score}
-                </p>
-                <p>
-                    Success Percentage {successRate}%
-                </p>
-                <p>Streak: {streak}</p>
-                <button onClick={this.props.resetScore}>Reset score</button>
-            </div>
+            <ul className="scorebar">
+                <li><strong>Level:</strong> {level}</li>
+                <li><strong>Questions Answered:</strong> {totalQuestions}</li>
+                <li><strong>Correct Answers:</strong> {score}</li>
+                <li><strong>Success Rate:</strong> {successRate}%</li>
+                <li><strong>Streak:</strong> {streak}</li>
+                <li><button onClick={this.props.resetScore}>Reset score</button></li>
+            </ul>
         )
     }
 }
