+++ src/components/frontpage.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 class FrontPage extends Component {
     render() { 
         return ( 
-            <div class="frontpage-wrapper">
+            <div className="frontpage-wrapper">
                 <h1>The dog breed app</h1>
                 <p>Test your knowledge about different dog breeds</p> 
             </div>
