+++ src/reducers/gameReducer.js
@@ -1,10 +1,11 @@
-import { SET_INITIAL_BREEDS, SET_CORRECT_ANSWER, SET_WRONG_ANSWERS } from '../actions/gameActions'
+import { SET_INITIAL_BREEDS, SET_CORRECT_ANSWER, SET_WRONG_ANSWERS, SHOW_BREED_HINT, RESET_HINT } from '../actions/gameActions'
 
 const initialState = {
     start: false,
     breeds: [],
     correctOption: null,
-    wrongOptions: []
+    wrongOptions: [],
+    hint: null
 }
 
 const reducer = (state = initialState, action = {}) => {
@@ -15,6 +16,10 @@ const reducer = (state = initialState, action = {}) => {
             return {...state, correctOption: {...action.payload}}
         case SET_WRONG_ANSWERS:
             return {...state, wrongOptions: [...action.payload]}
+        case SHOW_BREED_HINT:
+            return {...state, hint: `It's not a ${action.payload}`}
+        case RESET_HINT:
+            return {...state, hint: null}
         default:
             return state;
     }
